+++ src/common/utxobased/info/ecash.ts
@@ -80,7 +80,7 @@ export const coinInfo: CoinInfo = {
   includeCashaddrPrefix: true,
 
   prefixes: {
-    messagePrefix: ['\x18Bitcoin Signed Message::\n'],
+    messagePrefix: ['\x16eCash Signed Message:\n'],
     wif: [0x80],
     legacyXPriv: [0x0488ade4],
     legacyXPub: [0x0488b21e],